% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLibraryUsage.R
\name{addLibraryUsage}
\alias{addLibraryUsage}
\title{To add library annotation to a \linkS4class{GuideSet} object}
\usage{
addLibraryUsage(guideSet)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}
}
\value{
Adds a column for each possible library present in the
   \pkg{crisprAnnotation} package that indicated whether or not
   a given spacer sequence is fount in the given library.
}
\description{
Function to add annotation to a \linkS4class{GuideSet} object
   to describe whether or not a given spacer sequence is found in an 
   existing CRISPR gRNA library. \pkg{crisprAnnotation} package must be
   installed.
}
\examples{
\dontrun{
    guideSet <- addLibraryUsage(guideSetExample)
}
}
\author{
Luke Hoberecht
}
