% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSequenceFeatures.R
\name{addSequenceFeatures}
\alias{addSequenceFeatures}
\title{Add spacer sequence feature annotation columns to a
    \linkS4class{GuideSet} object}
\usage{
addSequenceFeatures(guideSet, addHairpin = FALSE, backbone = "AGGCTAGTCCGT")
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}

\item{addHairpin}{Should predicted hairpin formation via
sequence complementarity be calculated? FALSE by default.}

\item{backbone}{Backbone sequence in the guide RNA that is susceptible
to hairpin formation with a complementary region in the spacer sequence
(>= 5bp).}
}
\value{
\code{guideSet} with the following columns appended to
   \code{mcols(guideSet)}: 
   \itemize{
       \item \code{percentGC} — percent GC content
       \item \code{polyA}, \code{polyC}, \code{polyG}, \code{polyT} —
       presence of homopolymers of 4nt or longer
       \item \code{selfHairpin} — prediction of hairpin formation within the
       spacer sequence via self-complementarity (>= 4bp)
       \item \code{backboneHairpin} — prediction of hairpin formation with
       the backbone sequence via complementarity (>= 5bp)
   }
}
\description{
Add spacer sequence feature annotation columns, such as
    GC content, homopolymers, and hairpin predictions, to a
    \linkS4class{GuideSet} object.
}
\examples{
custom_seq <- c("ATTTCCGGAGGCGGAGAGGCGGGAGGAGCG")
data(SpCas9, package="crisprBase")
guideSet <- findSpacers(custom_seq, crisprNuclease=SpCas9)
guideSet <- addSequenceFeatures(guideSet)



}
