% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeSpacers.R
\name{convertToProtospacerGRanges}
\alias{convertToProtospacerGRanges}
\title{Convert PAM site coordinates to protospacer start and end coordinates}
\usage{
convertToProtospacerGRanges(guideSet)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}
}
\value{
A \linkS4class{GuideSet} object with start and end coordinates
    corresponding to the start and end coordinates of the protospacer
    sequences.
}
\description{
Convert PAM site coordinates to protospacer start
    and end coordinates.
}
\examples{
data(guideSetExample, package="crisprDesign")
gr <- convertToProtospacerGRanges(guideSetExample)

}
\author{
Jean-Philippe Fortin
}
