% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsDesign.R
\name{designOpsLibrary}
\alias{designOpsLibrary}
\title{Design gRNA library for optical pooled screening}
\usage{
designOpsLibrary(
  guideSet,
  n_cycles = 8,
  rt_direction = c("5prime", "3prime"),
  n_guides = 4,
  gene_field = "gene",
  min_dist_edit = 2,
  dist_method = c("hamming", "levenshtein"),
  splitByChunks = FALSE
)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}

\item{n_cycles}{Integer specifying the number of sequencing 
cycles used in the in situ sequencing. This effectively
determines the length of the barcodes to be used for
sequencing.}

\item{rt_direction}{String specifying from which direction
the reverse transcription of the gRNA spacer sequence
will occur. Must be either "5prime" or "3prime".
"5prime" by default.}

\item{n_guides}{Integer specifying how many gRNAs per
gene should be selected. 4 by default.}

\item{gene_field}{String specifying the column in \code{mcols(guideSet)}
specifying gene names.}

\item{min_dist_edit}{Integer specifying the minimum distance edit
required for barcodes to be considered dissimilar. Barcodes that
have edit distances less than the min_dist_edit will not be
included in the library. 2 by default.}

\item{dist_method}{String specifying distance method. 
Must be either "hamming" (default) or "levenshtein".}

\item{splitByChunks}{Should distances be calculated in a chunk-wise
manner? FALSE by default. Highly recommended when the set of query
barcodes is large to reduce memory footprint.}
}
\value{
A subset of the \code{GuideSet} containing the gRNAs
    selected for the OPS library.
}
\description{
Design gRNA library for optical pooled screening
}
\examples{
data(guideSetExample, package="crisprDesign")
guideSet <- unique(guideSetExample)
guideSet <- guideSet[1:200]

df <- data.frame(ID=names(guideSet),
                 spacer=spacers(guideSet, as.character=TRUE),
                 opsBarcode=as.character(guideSet$opsBarcode))

# Creating mock gene:
guideSet$gene <- rep(paste0("gene",1:10),each=20)
guideSet$rank <- rep(1:20,10)
opsLib1 <- designOpsLibrary(guideSet)
opsLib2 <- designOpsLibrary(guideSet, dist_method="levenshtein")

}
\author{
Jean-Philippe Fortin
}
