% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTssObjectFromTxObject.R
\name{getTssObjectFromTxObject}
\alias{getTssObjectFromTxObject}
\title{Extract TSS coordinates from a gene model object}
\usage{
getTssObjectFromTxObject(txObject)
}
\arguments{
\item{txObject}{A \linkS4class{TxDb} object or a \linkS4class{GRangesList}
object obtained using \code{\link{TxDb2GRangesList}} for annotating
on-target and off-target alignments using gene annotation.}
}
\value{
A GRanges object containing TSS coordinates
}
\description{
Extract TSS coordinates from a gene model object.
}
\examples{

data(grListExample, package="crisprDesign")
tss <- getTssObjectFromTxObject(grListExample)

}
\author{
Jean-Philippe Fortin
}
