% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grRepeatsExample}
\alias{grRepeatsExample}
\title{Example of a \linkS4class{GRanges} object containing repeat elements}
\format{
A \linkS4class{GRanges} object.
}
\usage{
data(grRepeatsExample, package="crisprDesign")
}
\description{
Example of a \linkS4class{GRanges} object containing genomic coordinates
    of repeat elements found in the neighborhood of human gene IQSEC3
    (ENSG00000120645).
}
\keyword{datasets}
