% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guideSetExample}
\alias{guideSetExample}
\title{Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
    the CDS of IQSEC3}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(guideSetExample, package="crisprDesign")
}
\description{
Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
   the coding sequence of human gene IQSEC3 (ENSG00000120645) for SpCas9
   nuclease.
}
\details{
The object was obtained by calling \code{\link{findSpacers}} on the
    CDS region of human gene IQSEC3. See code in
    \code{inst/scripts/generateGuideSet.R}.
}
\keyword{datasets}
