% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guideSetExampleFullAnnotation}
\alias{guideSetExampleFullAnnotation}
\title{Example of a fully-annotated \linkS4class{GuideSet} object storing gRNA
    sequences targeting the CDS of IQSEC3}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
guideSetExampleFullAnnotation
}
\description{
Example of a fully-annotated \linkS4class{GuideSet} object storing gRNA
   sequences targeting the coding sequence of human gene IQSEC3
   (ENSG00000120645) for SpCas9 nuclease.
}
\details{
The object was obtained by applying all available \code{add*}
    annotation functions in \code{crisprDesign} (e.g.
    \code{addSequenceFeatures}) to a randomly selected 20-guide subset of
     \code{guideSetExample}. See code in
    \code{inst/scripts/generateGuideSetFullAnnotation.R}.
}
\keyword{datasets}
