% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{guideSetExampleWithAlignments}
\alias{guideSetExampleWithAlignments}
\title{Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
    the CDS of IQSEC3 with off-target alignments.}
\format{
A \linkS4class{GuideSet} object.
}
\usage{
data(guideSetExampleWithAlignments, package="crisprDesign")
}
\description{
Example of a \linkS4class{GuideSet} object storing gRNA sequences targeting
   the coding sequence of human gene IQSEC3 (ENSG00000120645) for SpCas9
   nuclease with off-target alignments.
}
\details{
The object was obtained by adding off-target alignments
     to a randomly selected 20-guide subset of
     \code{guideSetExample}. See code in
    \code{inst/scripts/generateGuideSetFullAnnotation.R}.
}
\keyword{datasets}
