% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFC902.R
\name{pFC902insert}
\alias{pFC902insert}
\title{pFC902 insert design}
\usage{
pFC902insert(
  contig,
  siteofinterest,
  insert,
  direction,
  HMR1primers,
  HMR2primers,
  offset = 50,
  upstream = T,
  HMRlength = 1500,
  primersnumber = 5,
  insertpLength = 21,
  Paf_U3pLength = 23,
  gRNApLength = 23,
  tRNApLength = 21
)
}
\arguments{
\item{contig}{Sequence of contig that include site of interest}

\item{siteofinterest}{Insert site}

\item{insert}{Sequence of insert}

\item{direction}{+ or - (check if your insert is the same direction as your gene of interest for example)}

\item{HMR1primers}{A vector contains two sequences that contains the whole HMR1 region \
(I think it's easier to use NCBI primer finder for this to avoid specificity issues etc.)}

\item{HMR2primers}{A vector contains two sequences that contains the whole HMR2 region \
(I think it's easier to use NCBI primer finder for this to avoid specificity issues etc.)}

\item{offset}{Give some space between the two double strand break sites}

\item{upstream}{Find site upstream or downstream}

\item{primersnumber}{How many primers you want for double strand break}
}
\value{
a list of suggested primers, additional primers and primers required for pFC330~pFC333 cloning
}
\description{
design primers to introduce an insert into genome using pFC902
}
\details{
TBD
}
\examples{
genome<-seqinr::read.fasta("data/genome/GCA_021066465.1_ASM2106646v1_genomic.fna")
gff<-ape::read.gff("data/genome/GCA_021066465.1_ASM2106646v1_genomic.gff")
GoI<-"LI328DRAFT_134294"
genelocation<-gff[gff$type=="gene"&grepl(GoI,gff$attributes),]
contig<-genome[[as.character(genelocation$seqid)]]
siteofinterest<-ifelse(genelocation$strand=="+",genelocation$start,genelocation$end)
pFC902insert(contig,siteofinterest,pTCU1,"+",50) TBD
}
\references{
Nødvig, et al. 2018. Efficient oligo nucleotide mediated CRISPR-Cas9 gene editing in Aspergilli. Fungal Genet. Biol. 115:78-89.
}
