% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFC902.R
\docType{data}
\name{primer_CSN790}
\alias{primer_CSN790}
\title{primer CSN790}
\format{
An object of class \code{character} of length 1.
}
\usage{
primer_CSN790
}
\description{
primer CSN790
}
\keyword{datasets}
