% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFC902.R
\docType{data}
\name{promoter_Paf_U3}
\alias{promoter_Paf_U3}
\title{promoter Paf_U3}
\format{
An object of class \code{character} of length 1.
}
\usage{
promoter_Paf_U3
}
\description{
promoter Paf_U3
}
\keyword{datasets}
