% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\docType{data}
\name{promoter_pTCU1}
\alias{promoter_pTCU1}
\title{Common insert data}
\format{
An object of class \code{character} of length 1.
}
\usage{
promoter_pTCU1
}
\description{
Sequence for some common inserts
}
\details{
include pTCU1 from Trichoderma reesei
}
\keyword{datasets}
