% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTxObject.R
\name{queryTss}
\alias{queryTss}
\title{Convenience function to search for TSS coordinates.}
\usage{
queryTss(tssObject, queryColumn, queryValue, tss_window = NULL)
}
\arguments{
\item{tssObject}{A \linkS4class{GRanges} containing genomic positions
of transcription starting sites (TSSs).}

\item{queryColumn}{String specifying which column of \code{mcols(tssObject)}
should be searched for.}

\item{queryValue}{Character vector specifying the values to search for
in \code{tssObject[[queryColumn]]}.}

\item{tss_window}{Window size of genomic region flanking TSS to return.
Must be a numeric vector of length 2 consisting of the upstream limit
and downstream limit. Default is \code{c(-500, 500)}, which includes
500bp upstream and downstream of the TSS.}
}
\value{
A \linkS4class{GRanges} object. Searches yielding no results will
    return \code{NULL}.
}
\description{
Convenience function to search for TSS coordinates.
}
\seealso{
\code{\link{queryTxObject}} for querying gene annotations.
}
\author{
Luke Hoberecht, Jean-Philippe Fortin
}
