% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpObjectExample}
\alias{snpObjectExample}
\title{Example of a \linkS4class{GRanges} object containing SNPs}
\format{
A \linkS4class{GRanges} object.
}
\usage{
data(snpObjectExample, package="crisprDesign")
}
\description{
Example of a \linkS4class{GRanges} object containing genomic coordinates
    of SNPs in the neighborhood of human gene IQSEC3
    (ENSG00000120645).
}
\keyword{datasets}
