% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pFC902.R
\docType{data}
\name{terminator_Taf_U3}
\alias{terminator_Taf_U3}
\title{terminator Taf_U3}
\format{
An object of class \code{character} of length 1.
}
\usage{
terminator_Taf_U3
}
\description{
terminator Taf_U3
}
\keyword{datasets}
