% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tssObjectExample}
\alias{tssObjectExample}
\title{Example of a \linkS4class{GRanges} object containing TSS coordinates}
\format{
\linkS4class{GRanges} object of length 2 corresponding to the 2 
   TSSs of gene IQSEC3.
}
\usage{
tssObjectExample
}
\description{
Example of a \linkS4class{GRanges} containing transcription starting
    site (TSS) coordinates for human gene IQSEC3 (ENSG00000120645).
}
\details{
The TSS coordinates were obtained from the two transcript
    stored in the \code{grListExample} object for gene IQSEC3.
}
\keyword{datasets}
