% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{USERends}
\alias{USERends}
\title{USER site at ends}
\usage{
USERends(sequence)
}
\arguments{
\item{sequence}{Sequence of of interest}
}
\value{
the start and end of the two USER sites
}
\description{
Find sites Uracil Specific Excision Reaction at 5' and 3' ends of a sequence
}
\details{
TBD
}
\examples{
USERends(pTCU1)

}
