% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{USERPAM}
\alias{USERPAM}
\title{PAM site for USER}
\usage{
USERPAM(contig, siteofinterest, upstream = T, primersnumber = 5)
}
\arguments{
\item{contig}{Sequence of contig that include site of interest}

\item{siteofinterest}{Approximate site for double strand break}

\item{upstream}{Find site upstream or downstream}

\item{primersnumber}{How many PAMs you want}
}
\value{
exact
}
\description{
Find PAM sites with A and T upstream for Uracil Specific Excision Reaction
}
\details{
an internal function TBD
}
\examples{
USERPAM(contig,site-50)

}
