% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseEditing.R
\name{addEditedAlleles}
\alias{addEditedAlleles}
\title{To add edited alleles resulting from base editing}
\usage{
addEditedAlleles(txInfoDataFrame, substitution = "C2T")
}
\arguments{
\item{txInfoDataFrame}{A \code{DataFrame} obtained from \code{getTxInfoDataFrame}.}

\item{substitution}{String specifying the base editing substitution.
E.g: "C2T" to specify that Cs are converted to Ts.}
}
\value{
The original \code{txInfoDataFrame} object with the following
    additional columns:
        \code{aa_edited_fwd}: Amino acid resulting from editing forward strand
        \code{nuc_edited_fwd}: Nucleotide resulting from editing forward strand
        \code{class_fwd}: Type of mutation (silent, missense, or nonsense)
            resulting from editing forward strand
        \code{stop_fwd}: Does editing on the forward strand result in a stop codon?
        \code{aa_edited_rev}: Amino acid resulting from editing reverse strand
        \code{nuc_edited_rev}: Nucleotide resulting from editing reverse strand
        \code{class_rev}: Type of mutation (silent, missense, or nonsense)
            resulting from editing forward strand
        \code{stop_rev}: Does editing on the reverse strand result in a stop codon?
}
\description{
To add edited alleles resulting from base editing
}
\examples{

library(BSgenome.Hsapiens.UCSC.hg38)
bsgenome <- BSgenome.Hsapiens.UCSC.hg38
data("grListExample")
tx_id <- "ENST00000538872"
df <- getTxInfoDataFrame(tx_id=tx_id,
    txObject=grListExample,
    bsgenome=bsgenome)
df <- addEditedAlleles(df, substitution="C2T")

}
\author{
Jean-Philippe Fortin
}
