% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseEditing.R
\name{addEditedAlleles}
\alias{addEditedAlleles}
\title{To add edited alleles for a CRISPR base editing GuideSet}
\usage{
addEditedAlleles(
  guideSet,
  baseEditor,
  editingWindow = NULL,
  nMaxAlleles = 100,
  addFunctionalConsequence = TRUE,
  addSummary = TRUE,
  txTable = NULL,
  verbose = TRUE
)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}

\item{baseEditor}{A \linkS4class{BaseEditor} object.}

\item{editingWindow}{A numeric vector of length 2 specifying
start and end positions of the editing window with 
respect to the PAM site. If \code{NULL} (default),
the editing window of the \code{BaseEditor} object
will be considered.}

\item{nMaxAlleles}{Maximum number of edited alleles to report
for each gRNA. Alleles from high to low scores.
100 by default.}

\item{addFunctionalConsequence}{Should variant classification
of the edited alleles be added? TRUE by default.
If \code{TRUE}, \code{txTable} must be provided.}

\item{addSummary}{Should a summary of the variant classified
by added to the metadata columns of the \code{guideSet}
object? TRUE by default.}

\item{txTable}{Table of transcript-level nucleotide and amino
acid information needed for variant classification.
Usually returned by \code{\link{getTxInfoDataFrame}}.}

\item{verbose}{Should messages be printed to console?
TRUE by default.}
}
\value{
The original \code{guideSet} object with an additional
    metadata column (\code{editedAlleles}) storing the annotated
    edited alelles.
}
\description{
To add edited alleles for a CRISPR base editing GuideSet.
}
\examples{
\dontrun{
data(BE4max, package="crisprBase")
library(BSgenome.Hsapiens.UCSC.hg38)
bsgenome <- BSgenome.Hsapiens.UCSC.hg38
txid="ENST00000357654"
library(crisprDesignData)
txTable <- getTxInfoDataFrame(tx_id=txid,
    txObject=txdb_human,
    bsgenome=bsgenome)
gr <- queryTxObject(txdb_human,
    queryValue=txid,
    queryColumn="tx_id",
    featureType="cds")
gs <- findSpacers(gr,
    bsgenome=bsgenome,
    crisprNuclease=BE4max)
gs <- unique(gs)
gs <- gs[100:200]
gs <- addEditedAlleles(gs, baseEditor=BE4max,txTable=txTable)
}

}
\author{
Jean-Philippe Fortin
}
