% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOffTargetScores.R
\name{addOffTargetScores,GuideSet-method}
\alias{addOffTargetScores,GuideSet-method}
\alias{addOffTargetScores,PairedGuideSet-method}
\alias{addOffTargetScores,NULL-method}
\title{Add CFD and MIT scores to a \linkS4class{GuideSet} object.}
\usage{
\S4method{addOffTargetScores}{GuideSet}(object, max_mm = 2, includeDistance = TRUE, offset = 0)

\S4method{addOffTargetScores}{PairedGuideSet}(object, max_mm = 2, includeDistance = TRUE, offset = 0)

\S4method{addOffTargetScores}{`NULL`}(object)
}
\arguments{
\item{object}{A \linkS4class{GuideSet} object or a 
\linkS4class{PairedGuideSet} object.
\code{crisprNuclease(object)} must be either using SpCas9 or CasRx.}

\item{max_mm}{The maximimum number of mismatches between a spacer and
an off-target. Used to select which off-target scores to
include in calculating the aggregated score.}

\item{includeDistance}{Should a distance penalty for the MIT score be
included? TRUE by default.}

\item{offset}{Numeric value specifying an offset to add to the denominator
when calcuting the aggregated score (inverse summation formula).
0 by default.}
}
\value{
\code{guideSet} with off-target score columns appended in
    \code{mcols(alignments(guideSet))}. For SpCas9, it
    also adds aggregated off-target score columns to 
    \code{mcols(guideSet)}.
}
\description{
Add CFD and MIT off-target scores to a \linkS4class{GuideSet}
    object. Package \pkg{crisprScore} must be installed.
    Both CFD and MIT methods are available to SpCas9.
    CFD method is available for CasRx.
}
\details{
See \pkg{crisprScore} package for a description of the off-target
    scoring methods.
}
\examples{

data(guideSetExampleWithAlignments, package="crisprDesign")
gs <- guideSetExampleWithAlignments
gs <- addOffTargetScores(gs)

}
\seealso{
\code{link{addOnTargetScores}} to add on-target scores.
}
\author{
Jean-Philippe Fortin, Luke Hoberecht
}
