% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOffTargetScores.R
\name{addOffTargetScores}
\alias{addOffTargetScores}
\title{Add CFD and MIT scores to a \linkS4class{GuideSet} object.}
\usage{
addOffTargetScores(guideSet, max_mm = 2, includeDistance = TRUE, offset = 0)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}

\item{max_mm}{The maximimum number of mismatches between a spacer and
an off-target. Used to select which off-target scores to
include in calculating the aggregated score.}

\item{includeDistance}{Should a distance penalty for the MIT score be
included? TRUE by default.}

\item{offset}{Numeric value specifying an offset to add to the denominator
when calcuting the aggregated score (inverse summation formula). 0 by
default.}
}
\value{
\code{guideSet} with off-target score columns appended in
    \code{mcols(alignments(guideSet))} and aggregated off-target score
    columns appended in \code{mcols(guideSet)}.
}
\description{
Add CFD and MIT off-target scores to a \linkS4class{GuideSet}
    object. Package \pkg{crisprScore} must be installed.
}
\details{
See \pkg{crisprScore} package for a description of the off-target
    scoring methods.
}
\examples{

# Creating a bowtie index:
library(Rbowtie)
fasta <- system.file(package="crisprDesign", "fasta/chr12.fa")
outdir <- tempdir()
Rbowtie::bowtie_build(fasta,
                      outdir=outdir,
                      force=TRUE,
                      prefix="chr12")
bowtieIndex <- file.path(outdir, "chr12")

# Adding spacer alignments:
data(guideSetExample, package="crisprDesign")
guideSetExample <- guideSetExample[1:3]
guideSet <- addSpacerAlignments(guideSetExample,
                                bowtie_index=bowtieIndex,
                                n_mismatches=1,
                                txObject=grListExample)

# Adding off-target scores:
guideSet <- addOffTargetScores(guideSet)

}
\seealso{
\code{link{addOnTargetScores}} to add on-target scores.
}
\author{
Jean-Philippe Fortin, Luke Hoberecht
}
