% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addOnTargetScores.R
\name{addOnTargetScores}
\alias{addOnTargetScores}
\title{Add on-target scores to a \linkS4class{GuideSet} object.}
\usage{
addOnTargetScores(
  guideSet,
  enzyme = c("WT", "ESP", "HF"),
  promoter = c("U6", "T7"),
  methods = c("azimuth", "deephf", "ruleset1", "lindel", "deepcpf1", "enpamgb",
    "crisprscan")
)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}

\item{enzyme}{Character string specifying the Cas9 variant to be used
for DeepHF scoring. Wildtype Cas9 (WT) by default. See details below.}

\item{promoter}{Character string speciyfing promoter used for expressing 
sgRNAs for wildtype Cas9 (must be either "U6" or "T7") for DeepHF
scoring. "U6" by default.}

\item{methods}{Character vector specifying method names for on-target
efficiency prediction algorithms.}
}
\value{
\code{guideSet} with columns of on-target scores appended in
    \code{mcols(guideSet)}.
}
\description{
Add on-target scores to a \linkS4class{GuideSet} object
   for all methods available in the \pkg{crisprScore} package for a given
   CRISPR nuclease. Requires \pkg{crisprScore} package to be installed.
}
\details{
See \pkg{crisprScore} package for a description of each score.
}
\examples{
if (interactive()){
    gs <- findSpacers("CCAACATAGTGAAACCACGTCTCTATAAAGAATACAAAAAATTAGCCGGGTTA")
    gs <- addOnTargetScores(gs)
}

}
\seealso{
\code{\link{addOffTargetScores}} to add off-target scores.
}
\author{
Jean-Philippe Fortin, Luke Hoberecht
}
