% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/addSequenceFeatures.R
\name{addSequenceFeatures}
\alias{addSequenceFeatures}
\alias{addSequenceFeatures,GuideSet-method}
\alias{addSequenceFeatures,PairedGuideSet-method}
\alias{addSequenceFeatures,NULL-method}
\title{Add spacer sequence feature annotation columns to a
    \linkS4class{GuideSet} object}
\usage{
addSequenceFeatures(object, ...)

\S4method{addSequenceFeatures}{GuideSet}(
  object,
  addHairpin = FALSE,
  backbone = "AGGCTAGTCCGT",
  tp53 = TRUE,
  ...
)

\S4method{addSequenceFeatures}{PairedGuideSet}(
  object,
  addHairpin = FALSE,
  backbone = "AGGCTAGTCCGT",
  tp53 = TRUE,
  ...
)

\S4method{addSequenceFeatures}{`NULL`}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{GuideSet} or a 
\linkS4class{PairedGuideSet} object.}

\item{...}{Additional arguments, currently ignored.}

\item{addHairpin}{Whether to include predicted hairpin formation via
sequence complementarity. FALSE by default. See details.}

\item{backbone}{Backbone sequence in the guide RNA that is susceptible
to hairpin formation with a complementary region in the spacer sequence.}

\item{tp53}{Should TP53-related toxicity features be added? 
TRUE by default.}
}
\value{
\code{guideSet} with the following columns appended to
   \code{mcols(guideSet)}: 
   \itemize{
       \item \code{percentGC} — percent GC content
       \item \code{polyA}, \code{polyC}, \code{polyG}, \code{polyT} —
       presence of homopolymers of 4nt or longer
       \item \code{selfHairpin} — prediction of hairpin formation within the
       spacer sequence via self-complementarity
       \item \code{backboneHairpin} — prediction of hairpin formation with
       the backbone sequence via complementarity
   }
}
\description{
Add spacer sequence feature annotation columns, such as
    GC content, homopolymers, and hairpin predictions, to a
    \linkS4class{GuideSet} object.
}
\details{
\code{addSequenceFeatures} predicts spacers to form internal
    hairpins when there is a palindromic sequence within the spacer having
    arms of >=4nt and >=50\% GC content, and are separated by a loop of
    >=4nt. Backbone hairpin formation is predicted when the spacer and
    backbone share a complementary sequence of >=5nt and >=50\% GC content.
}
\examples{
custom_seq <- c("ATTTCCGGAGGCGGAGAGGCGGGAGGAGCG")
data(SpCas9, package="crisprBase")
guideSet <- findSpacers(custom_seq, crisprNuclease=SpCas9)
guideSet <- addSequenceFeatures(guideSet)


}
