% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsDesign.R
\name{designOpsLibrary}
\alias{designOpsLibrary}
\title{Design gRNA library for optical pooled screening}
\usage{
designOpsLibrary(
  df,
  n_guides = 4,
  gene_field = "gene",
  min_dist_edit = 3,
  dist_method = c("hamming", "levenstein")
)
}
\arguments{
\item{df}{data.frame containing information about candidate
gRNAs from which to build the OPS library. See details.}

\item{n_guides}{Integer specifying how many gRNAs per
gene should be selected. 4 by default.}

\item{gene_field}{String specifying the column in \code{df}
specifying gene names.}

\item{min_dist_edit}{Integer specifying the minimum distance edit
required for barcodes to be considered dissimilar. Barcodes that
have edit distances less than the min_dist_edit will not be
included in the library.}

\item{dist_method}{String specifying distance method. Must be
either "hamming" (default) or "levenstein".}
}
\description{
Design gRNA library for optical pooled screening
}
\examples{
data("guideSetExample")
guideSet <- guideSetExample
guideSet <- addOpsBarcodes(guideSet)
df <- data.frame(ID=names(guideSet),
                 spacer=spacers(guideSet, as.character=TRUE),
                 opsBarcode=as.character(guideSet$opsBarcode))
df$gene <- rep(paste0("gene",1:40),each=20)
df$rank <- rep(1:20,40)
opsLib <- designOpsLibrary(df)
}
\author{
Jean-Philippe Fortin
}
