% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSpacers.R
\name{findSpacers}
\alias{findSpacers}
\title{Find CRISPR gRNA spacer sequences in a set of DNA sequences.}
\usage{
findSpacers(
  x,
  bsgenome = NULL,
  crisprNuclease = NULL,
  canonical = TRUE,
  spacer_len = NULL,
  both_strands = TRUE,
  cut_offset = NULL,
  strict_overlap = TRUE,
  collapse = TRUE,
  remove_ambiguities = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{Either a \linkS4class{GRanges} object or a character vector of
genomic sequences.}

\item{bsgenome}{\linkS4class{BSgenome} object from which to extract
sequences if a \linkS4class{GRanges} object is provided as input.}

\item{crisprNuclease}{A \linkS4class{CrisprNuclease} object.}

\item{canonical}{Whether to return only guide sequences having canonical
PAM sequences. If TRUE (default), only PAM sequences with the highest
weights stored in the \code{crisprNuclease} object will be considered.}

\item{spacer_len}{Length of spacer to be used if different from
default length specified in \code{crisprNuclease}.}

\item{both_strands}{Should both strands be considered to search for 
protospacer sequences?
TRUE by default.}

\item{cut_offset}{Distance in nucleotides between \code{pam_site}
and \code{cut_site} if different from default offset specified by
\code{crisprNuclease}.}

\item{strict_overlap}{If \code{TRUE} (default), only include guides
having the cut site in the input range. Ignored when input is a
character vector of genomic sequences.}

\item{collapse}{For input containing multiple regions, whether to return
only unique results:collapsing results will remove duplicated guides
found in overlapping ranges/regions.}

\item{remove_ambiguities}{Should spacer sequences with ambiguous 
nucleotides ("N") be removed?
TRUE by default.}

\item{verbose}{Should messages be printed to console?
TRUE by default.}
}
\value{
A \linkS4class{GuideSet} object.
}
\description{
Returns all possible, valid gRNA sequences from either a
   \linkS4class{GRanges} object or a character vector of genomic sequences,
   for a given CRISPR nuclease.
}
\details{
For input that is a character vector of genomic sequences,
    all identified guides are returned with their relative strand and
    coordinates. The input sequence is assumed to be on the positive
    strand starting at position 1.
}
\examples{
# Using custom sequence as input:
guides <- findSpacers("CCAANAGTGAAACCACGTCTCTATAAAGAATACAAAAAATTAGCCGGGTGTTA")

# Exon-intro region of human KRAS specified 
# using a GRanges object:
if (requireNamespace("BSgenome.Hsapiens.UCSC.hg38")){
    library(GenomicRanges)
    bsgenome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38

    gr_input <- GRanges(c("chr12"),
                        IRanges(start=25224014, end=25227007))
    genome(gr_input) <- "hg38"
    guideSet <- findSpacers(gr_input, bsgenome=bsgenome)

    # Designing guides for enAsCas12a nuclease:
    data(enAsCas12a, package="crisprBase")
    guideSet <- findSpacers(gr_input, 
                            canonical=FALSE,
                            bsgenome=bsgenome,
                            crisprNuclease=enAsCas12a)
}

}
\author{
Jean-Philippe Fortin, Luke Hoberecht
}
