% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMrnaSequence.R
\name{getMrnaSequences}
\alias{getMrnaSequences}
\title{To obtain mRNA nucleotide sequences for a transcript}
\usage{
getMrnaSequences(txids, txObject, bsgenome)
}
\arguments{
\item{txids}{Character vector specifying Ensembl transcript IDs.}

\item{txObject}{A \linkS4class{TxDb} object or a \linkS4class{GRangesList}
object obtained using \code{\link{TxDb2GRangesList}} for annotating
on-target and off-target alignments using gene annotation.}

\item{bsgenome}{A \linkS4class{BSgenome} object from which to extract
sequences if \code{x} is a \linkS4class{GRanges} object.}
}
\value{
DNAStringSet object representing mRNA sequences.
}
\description{
To obtain mRNA nucleotide sequences for a transcript.
}
\examples{
\dontrun{
library(crisprDesignData)
library(BSgenome.Hsapiens.UCSC.hg38)
bsgenome <- BSgenome.Hsapiens.UCSC.hg38
txObject <- txdb_human
txids <- c("ENST00000457313",
"ENST00000256078")

out <- getMrnaSequences(txids,
    bsgenome=bsgenome,
    txObject=txObject)
}

}
\author{
Jean-Philippe Fortin
}
