% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTxInfoDataFrame.R
\name{getTxInfoDataFrame}
\alias{getTxInfoDataFrame}
\title{To obtain a DataFrame of transcript-specific CDS and mRNA coordinates}
\usage{
getTxInfoDataFrame(
  tx_id,
  txObject,
  bsgenome,
  extend = 30,
  checkCdsLength = TRUE
)
}
\arguments{
\item{tx_id}{String specifying ENSEMBL Transcript id.}

\item{txObject}{A \linkS4class{TxDb} object or a \linkS4class{GRangesList}
object obtained using \code{\link{TxDb2GRangesList}}.}

\item{bsgenome}{\linkS4class{BSgenome} object from which to extract
sequences if a \linkS4class{GRanges} object is provided as input.}

\item{extend}{Integer value specifying how many nucleotides in intron regions
should be included.}

\item{checkCdsLength}{Should the CDS nucleotide length be a multiple of 3?
TRUE by default.}
}
\value{
A \code{DataFrame} containing nucleotide and amino acid information.
}
\description{
To obtain a DataFrame of transcript-specific CDS and mRNA coordinates.
}
\examples{

library(BSgenome.Hsapiens.UCSC.hg38)
bsgenome <- BSgenome.Hsapiens.UCSC.hg38
data("grListExample")
tx_id="ENST00000538872"
df <- getTxInfoDataFrame(tx_id=tx_id,
    txObject=grListExample,
    bsgenome=bsgenome)

}
\author{
Jean-Philippe Fortin
}
