% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{seqnames}
\alias{mcols}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{GenomeInfoDb}{\code{\link[GenomeInfoDb:seqinfo]{seqnames}}}

  \item{S4Vectors}{\code{\link[S4Vectors:Vector-class]{mcols}}}
}}

