% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addRepeats.R
\name{removeRepeats}
\alias{removeRepeats}
\title{Remove \linkS4class{GuideSet} gRNAs that overlap repeat elements}
\usage{
removeRepeats(guideSet, gr.repeats, ignore.strand = TRUE)
}
\arguments{
\item{guideSet}{A \linkS4class{GuideSet} object.}

\item{gr.repeats}{A \linkS4class{GRanges} object containing
repeat elements regions.}

\item{ignore.strand}{Should gene strand be ignored when annotating?
TRUE by default.}
}
\value{
\code{guideSet} filtered for spacer sequences not overlapping
    any repeat elements. An \code{inRepeats} column is also appended in
    \code{mcols(guideSet)}.
}
\description{
Remove \linkS4class{GuideSet} gRNAs that overlap repeat
    elements.
}
\examples{
data(guideSetExample, package="crisprDesign")
data(grRepeatsExample, package="crisprDesign")
guideSet <- removeRepeats(guideSetExample,
                          gr.repeats=grRepeatsExample)

}
\seealso{
\code{link{addRepeats}}.
}
\author{
Jean-Philippe Fortin, Luke Hoberecht
}
